fix(acls): UserCache.delete/1 breaking change
TD-4212 (#91) changed the type of `UserCache.delete_user/1` making it
incompatible with the public API `UserCache.delete/1` which expects to
receive a user_id as the argument.

This commit restores UserCache to it's previous behaviour. Instead of
deleting `acl_role_users` entries when a user is deleted, AclCache now
uses Redis `SINTER` to intersect the ACL's user ids with all current
user ids.